Fit the single species, single season site occupancy model.

This function fits the single species, single season site occupancy model first developed by MacKenzie et al. (2002).

occ_mod(
  occupancy,
  detection,
  data,
  niter = 1000,
  nchains = 3,
  seed = NULL,
  save_model = FALSE,
  model_name = paste0("occ_model_", Sys.Date()),
  beta_prior = "dnorm(0, 1/2)"
)

Arguments

occupancy
model declaration for the occupancy portion of the model using standard linear model syntax

detection
model declaration for the detection portion of the model using standard linear model syntax

data
data from which to create the detection and covariate matrices. Each row should represent a single visit to a site. See details and examples for proper formatting

niter
number of MCMC iterations

nchains
number of MCMC chains

seed
optional seed

save_model
logical; should a text file containing the model be exported?

model_name
character string defining the name of the text file describing the model if save_model = TRUE

beta_prior
character string defining prior distribution for regression coefficients at the occupancy and detection levels. Priors should be specified using distributions available in NIMBLE. See available distributons in NIMBLE

Value

an object of class list containing the following:

samples object of class list of length nchains, each containing a matrix posterior samples

loglik object of class list of length codenchains, each containing a matrix of samples of the log posterior likelihood

Details

This function fits the single season, single species site occupancy model using the logit link function. The data should contain columns named site, visit, and y. See examples.

Examples

# simulate data
sim <- sim_occ(M = 100, max_j = 10, seed = 01012021, rand_visits = FALSE)
data <- sim$data

# note structure of data frame
names(data)

#> [1] "site"     "visit"    "y"        "psi_cov1" "p_cov1"  

head(data)

#>   site visit y   psi_cov1     p_cov1
#> 1    1     1 0 0.07477433 0.05875483
#> 2    1     2 0 0.07477433 1.01146648
#> 3    1     3 0 0.07477433 0.53167259
#> 4    1     4 0 0.07477433 0.18080062
#> 5    1     5 0 0.07477433 1.08585510
#> 6    1     6 0 0.07477433 1.47234856

# fit model
ex <- occ_mod(occupancy = ~ psi_cov1, detection = ~ p_cov1, data = data,
              niter = 4000, beta_prior = "dunif(-5, -2)")

#> # priors
#> for(i in 1:2){beta_psi[i] ~ dunif(-5, -2)}
#> for(i in 1:2){beta_p[i] ~ dunif(-5, -2)}
#> 
#> # likelihood 
#> # loop through sites 
#> for(i in 1:M){ 
#>  # calculate psi 
#>  logit(psi[i]) <- beta_psi[1] +
#>   beta_psi[2] * psi_cov1[i]
#> 
#>  # latent occupancy state 
#>  z[i] ~ dbern(psi[i]) 
#> 
#>  # loop through visits 
#>  for(j in 1:nvisits[i]){
#>   # calculate p
#>   logit(p[i,j]) <- beta_p[1] +
#>    beta_p[2] * p_cov1[i,j]
#> 
#>   # response model
#>   Y[i,j] ~ dbern(z[i] * p[i,j])
#>  }
#> }

#> 
#> Building R model

#> defining model...

#> Detected Y as data within 'constants'.

#> Adding Y as data for building model.

#> building model...

#> setting data and initial values...

#> running calculate on model (any error reports that follow may simply reflect missing values in model variables) ... 

#> 

#> checking model sizes and dimensions...

#> 

#> model building finished.

#> 
#> Compiling R model to C

#> compiling... this may take a minute. Use 'showCompilerOutput = TRUE' to see C++ compilation details.

#> compilation finished.

#> 
#> Building R MCMC object

#> 
#> Compiling R MCMC object to C

#> compiling... this may take a minute. Use 'showCompilerOutput = TRUE' to see C++ compilation details.

#> compilation finished.

#> 
#> Running MCMC

#> running chain 1...

#> warning: problem initializing stochastic node beta_psi[1]: logProb is -Inf.
#> warning: problem initializing stochastic node beta_psi[2]: logProb is -Inf.
#> warning: problem initializing stochastic node beta_p[1]: logProb is -Inf.
#> warning: problem initializing stochastic node beta_p[2]: logProb is -Inf.
#> |-------------|-------------|-------------|-------------|
#> |-------------------------------------------------------|

#> running chain 2...

#> warning: problem initializing stochastic node beta_psi[1]: logProb is -Inf.
#> warning: problem initializing stochastic node beta_psi[2]: logProb is -Inf.
#> warning: problem initializing stochastic node beta_p[1]: logProb is -Inf.
#> warning: problem initializing stochastic node beta_p[2]: logProb is -Inf.
#> |-------------|-------------|-------------|-------------|
#> |-------------------------------------------------------|

#> running chain 3...

#> warning: problem initializing stochastic node beta_psi[1]: logProb is -Inf.
#> warning: problem initializing stochastic node beta_psi[2]: logProb is -Inf.
#> warning: problem initializing stochastic node beta_p[1]: logProb is -Inf.
#> warning: problem initializing stochastic node beta_p[2]: logProb is -Inf.
#> |-------------|-------------|-------------|-------------|
#> |-------------------------------------------------------|

#> 
#> Computing posterior likelihood

#> 
#> Done.

#> 
#> Computing fitted values.

#> 
#> Done.

# results
ex

#> Single species, single season site occupancy model fit using NIMBLE.
#> 
#> The following model was fit:
#> # priors
#> for(i in 1:2){beta_psi[i] ~ dunif(-5, -2)}
#> for(i in 1:2){beta_p[i] ~ dunif(-5, -2)}
#> 
#> # likelihood 
#> # loop through sites 
#> for(i in 1:M){ 
#>  # calculate psi 
#>  logit(psi[i]) <- beta_psi[1] +
#>   beta_psi[2] * psi_cov1[i]
#> 
#>  # latent occupancy state 
#>  z[i] ~ dbern(psi[i]) 
#> 
#>  # loop through visits 
#>  for(j in 1:nvisits[i]){
#>   # calculate p
#>   logit(p[i,j]) <- beta_p[1] +
#>    beta_p[2] * p_cov1[i,j]
#> 
#>   # response model
#>   Y[i,j] ~ dbern(z[i] * p[i,j])
#>  }
#> }
#> 
#> The following MCMC algorithm was used:
#> ===== Monitors =====
#>  thin = 1: z, beta_psi, beta_p
#>  ===== Samplers =====
#>  RW sampler (4)
#>    - beta_psi[]  (2 elements)
#>    - beta_p[]  (2 elements)
#>  binary sampler (100)
#>    - z[]  (100 elements)

summary(ex)

#> Call:
#> occ_mod(occupancy = ~psi_cov1, detection = ~p_cov1)
#> 
#> MCMC information:
#> nchains: 3
#> niter: 4000
#> burnin: 2000
#> 
#> Multivariate potential scale reduction factor:
#>  860.2375
#> 
#> Potential scale reduction factors:
#>             Point est. Upper C.I.
#> beta_psi[1] 179.1       1119     
#> beta_psi[2] 115.1        307     
#> beta_p[1]   912.5       2334     
#> beta_p[2]   922.4       2708     
#> 
#> Occupancy coefficients:
#>             estimate sd    0.025  0.5    0.975 n_eff
#> beta_psi[1] 0.2014   1.583 -2.078 1.166  1.469 470.9
#> beta_psi[2] -0.8423  1.684 -2.109 -2.009 1.538 75.26
#> 
#> Detection coefficients:
#>           estimate sd    0.025  0.5    0.975  n_eff
#> beta_p[1] -1.123   1.245 -2.011 -2.001 0.6376  1585
#> beta_p[2] -0.955   1.483 -2.012 -2.001 1.142   1265
#> 
#> Type 2 waic: 7512.8948

str(ex)

#> List of 3
#>  $ samples:List of 3
#>   ..$ chain1: num [1:4000, 1:104] -0.213 -0.213 -0.213 -0.213 -0.213 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:104] "beta_p[1]" "beta_p[2]" "beta_psi[1]" "beta_psi[2]" ...
#>   ..$ chain2: num [1:4000, 1:104] 0.638 0.638 0.638 0.638 0.638 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:104] "beta_p[1]" "beta_p[2]" "beta_psi[1]" "beta_psi[2]" ...
#>   ..$ chain3: num [1:4000, 1:104] -0.513 -0.513 -0.513 -0.513 -0.513 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:104] "beta_p[1]" "beta_p[2]" "beta_psi[1]" "beta_psi[2]" ...
#>  $ loglik :List of 3
#>   ..$ chain1: num [1:4000, 1:100] -1.42 -1.42 -1.42 -1.42 -1.42 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:100] "site1" "site2" "site3" "site4" ...
#>   ..$ chain2: num [1:4000, 1:100] -1.73 -1.73 -1.73 -1.73 -1.73 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:100] "site1" "site2" "site3" "site4" ...
#>   ..$ chain3: num [1:4000, 1:100] -0.3054 -0.0905 -0.0905 -0.0905 -0.1197 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:100] "site1" "site2" "site3" "site4" ...
#>  $ derived:List of 2
#>   ..$ psi:List of 3
#>   .. ..$ chain1: num [1:4000, 1:100] 0.758 0.758 0.758 0.758 0.758 ...
#>   .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. ..$ : NULL
#>   .. .. .. ..$ : chr [1:100] "site1" "site2" "site3" "site4" ...
#>   .. ..$ chain2: num [1:4000, 1:100] 0.822 0.822 0.822 0.822 0.822 ...
#>   .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. ..$ : NULL
#>   .. .. .. ..$ : chr [1:100] "site1" "site2" "site3" "site4" ...
#>   .. ..$ chain3: num [1:4000, 1:100] 0.2631 0.0865 0.0865 0.0865 0.1128 ...
#>   .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. ..$ : NULL
#>   .. .. .. ..$ : chr [1:100] "site1" "site2" "site3" "site4" ...
#>   ..$ p  :List of 3
#>   .. ..$ chain1: num [1:4000, 1:100, 1:10] 0.437 0.437 0.437 0.437 0.437 ...
#>   .. .. ..- attr(*, "dimnames")=List of 3
#>   .. .. .. ..$ : NULL
#>   .. .. .. ..$ : chr [1:100] "site1" "site2" "site3" "site4" ...
#>   .. .. .. ..$ : chr [1:10] "visit1" "visit2" "visit3" "visit4" ...
#>   .. ..$ chain2: num [1:4000, 1:100, 1:10] 0.658 0.658 0.658 0.658 0.658 ...
#>   .. .. ..- attr(*, "dimnames")=List of 3
#>   .. .. .. ..$ : NULL
#>   .. .. .. ..$ : chr [1:100] "site1" "site2" "site3" "site4" ...
#>   .. .. .. ..$ : chr [1:10] "visit1" "visit2" "visit3" "visit4" ...
#>   .. ..$ chain3: num [1:4000, 1:100, 1:10] 0.39 0.39 0.39 0.39 0.39 ...
#>   .. .. ..- attr(*, "dimnames")=List of 3
#>   .. .. .. ..$ : NULL
#>   .. .. .. ..$ : chr [1:100] "site1" "site2" "site3" "site4" ...
#>   .. .. .. ..$ : chr [1:10] "visit1" "visit2" "visit3" "visit4" ...
#>  - attr(*, "class")= chr [1:2] "occ_mod" "list"
#>  - attr(*, "code")= chr "# priors\nfor(i in 1:2){beta_psi[i] ~ dunif(-5, -2)}\nfor(i in 1:2){beta_p[i] ~ dunif(-5, -2)}\n\n# likelihood "| __truncated__
#>  - attr(*, "mcmcinfo")= chr [1:8] "===== Monitors =====" "thin = 1: z, beta_psi, beta_p" "===== Samplers =====" "RW sampler (4)" ...
#>  - attr(*, "occupancy_call")=Class 'formula'  language ~psi_cov1
#>   .. ..- attr(*, ".Environment")=<environment: 0x7fb80273e840> 
#>  - attr(*, "detection_call")=Class 'formula'  language ~p_cov1
#>   .. ..- attr(*, ".Environment")=<environment: 0x7fb80273e840>