Simulate data from the multi-species, misclassification count model for a single season.

This function simulates data from he multi-species, misclassification count occupancy model first developed by Wright et al. (2020).

sim_MSMocc(
  K = 2,
  M = 20,
  M.conf = 2,
  max_j = 10,
  beta_psi = matrix(c(0, 1, 0, 1.2), ncol = K),
  beta_lambda = matrix(c(0, 1, 0, 1.2), ncol = K),
  theta = matrix(c(0.8, 0.2, 0.2, 0.8), byrow = T, ncol = K),
  seed = NULL,
  rand_visits = TRUE
)

Arguments

K
number of species

M
number of sites

M.conf
number of sites that have confirmed observations

max_j
maximum number of visits to each site. If rand_visits = FALSE, this value is the number of visits to each site. If rand_visits = TRUE, each site is visited a random (sample(2:max_j, size = 1)) number of times.

beta_psi
vector of regression coefficients used to generate psi

beta_lambda
vector of regression coefficients used to generate lambda

theta
confusion matrix for automatic call classifications. Rows of theta should sum to one, diagonal elements are correct identification probabilities, first row corresponds to probabilities for calls truly from species 1.

seed
optional seed for reproducibility

rand_visits
logical; should each site be visited a random number of times? See details.

Value

object of class list containing the following elements:

beta_psi vector of regression coefficients used to generate psi

beta_lambda vector of regression coefficients used to generate lambda

psi_cov matrix of site level covariates

lambda_cov array of detection level covariates; each slice represents a single covariate

psi vector of derived site level occupancy probabilities

lambda matrix of derived visit level expected count detection parameter

z matrix of latent occupancy states for each site/species

Y array of observed Bernoulli responses for site/visit/species

n_visits vector of number of visits to each site

c.conf how calls confirmed to species K were classified for confirmed visits

c.sum.a and c.sum.b unconfirmed calls classified TO species K from sites with some confirmations ("a") or no confirmations ("b")

sites.conf site locations with confirmations

sites.unconf.a and sites.unconf.b site locations without confirmations ("a" indecates some observations at these sites were confirmed, "b" sites had no confirmations)

data a data frame containing all information necessary to fit the model

Details

This function simulates data from the single season, multiple species occupancy model to estimate occupancy and relative activity for multiple species simultaneously. If rand_visits = TRUE, each site is visited a random number of times between two and max_j. Covariates are drawn from the uniform(0, 1.5) distribution so that the effect of the direction of each regression coefficient is intuitive. Note that if covariates are not desired, beta_psi and beta_lambda can be set to intercepts that generate the desired derived probabilities.

Examples

n.species <- 4
sim <- sim_MSMocc(K = n.species, M = 100, M.conf = 10, max_j = 20,
                  beta_psi = matrix(c(0, 1,
                                      0, 1.1,
                                      0, 1.2,
                                      0, 1.3),
                                    ncol = n.species),
                  beta_lambda = matrix(c(0, 1, 1,
                                         0, 1, 1,
                                         0, 1, 1.2,
                                         0, 1, 1.4),
                                       ncol = n.species),
                  theta = matrix(c(0.7, 0.1, 0.1, 0.1,
                            0.1, 0.7, 0.1, 0.1,
                            0.1, 0.1, 0.7, 0.1,
                            0.1, 0.1, 0.1, 0.7), nrow = n.species),
                  rand_visits = TRUE)
sim$Y

#> , , 1
#> 
#>        [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>   [1,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>   [2,]    1    2    3    1   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>   [3,]    1    1    0    1    1    2    1    0    2     1     0     0     1
#>   [4,]    0    0    1    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>   [5,]    0    0    0    0    0    0    0    0   NA    NA    NA    NA    NA
#>   [6,]    0    0    0    0    0    0    0    0    0     0     0    NA    NA
#>   [7,]    0    0    0    0    0    0    0   NA   NA    NA    NA    NA    NA
#>   [8,]    0    0    0    0    0    0    0    0    0     0     0    NA    NA
#>   [9,]    0    0    0    0    0    0    0    0    0     0     0     0    NA
#>  [10,]    0    0    1    0    1    1    1    0    0     2     0     0     1
#>  [11,]    0    0    0    0    0    0    0    0    0     0     0    NA    NA
#>  [12,]    0    0    0    1    1    0    0    0    2     0    NA    NA    NA
#>  [13,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [14,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [15,]    1    0    0    0    0    1    0    0    2     2     3     0     1
#>  [16,]    0    0    1    0    2    0    0    1    1     1     0     0     0
#>  [17,]    0    2    1    1    0    0    0    0    1     0     2     3     0
#>  [18,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [19,]    2    1    3    2    0    1    2    1    3     1     0     1     0
#>  [20,]    1    1    1    1    0    0   NA   NA   NA    NA    NA    NA    NA
#>  [21,]    0    0    0    1    3    0    1    1    0     1     1     0     2
#>  [22,]    0    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [23,]    1    1    1    2    2    0    0    2    0     1     1     0     0
#>  [24,]    0    2    0    1    0    3   NA   NA   NA    NA    NA    NA    NA
#>  [25,]    2    0    2    0    0    0    2    1   NA    NA    NA    NA    NA
#>  [26,]    0    1    0    1    3    2    0    2    2     1     0     0     0
#>  [27,]    1    1    1    1    0   NA   NA   NA   NA    NA    NA    NA    NA
#>  [28,]    1    0    2    1    1    0    0    0    0     2    NA    NA    NA
#>  [29,]    0    3    1    0    1   NA   NA   NA   NA    NA    NA    NA    NA
#>  [30,]    0    0    0    0    0    0    0    0   NA    NA    NA    NA    NA
#>  [31,]    0    1    0    0    0    2    0    0    0    NA    NA    NA    NA
#>  [32,]    3    0    0    1    1    2    3    0    1     0     0     0     0
#>  [33,]    0    2    1    1    0    1    0    2    1     1     1    NA    NA
#>  [34,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [35,]    1    1    0    1    3    0    1    0    1    NA    NA    NA    NA
#>  [36,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [37,]    0    1    1    0    2    0    3    3    1     0     1     1     1
#>  [38,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [39,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [40,]    0    0    1    0    0    1    0    2    1     0     1     1     3
#>  [41,]    1    1    1    4    2    2    0    0    2     0     1     1     0
#>  [42,]    0    1    2    1   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [43,]    0    0    1    2    1    1    1    0    3     1     0     1     0
#>  [44,]    1    1    1    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [45,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [46,]    0    0    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [47,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [48,]    0    1    1   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [49,]    1    1    0    0    0    1    0    1   NA    NA    NA    NA    NA
#>  [50,]    0    0    1    0    1    0    2    0    1     2     1     1     1
#>  [51,]    0    1    0    3   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [52,]    0    2    1    0    0    3    0    1    1     1     0     3     0
#>  [53,]    2    2   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [54,]    1    1    0    0    0    0    2    0    1     0     0     0    NA
#>  [55,]    1    1    0    1    2    1    2    1    1     0    NA    NA    NA
#>  [56,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [57,]    0    0    0    0    0    0    0    0    0     0    NA    NA    NA
#>  [58,]    1    0    0    3   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [59,]    1    0    1    1    0    1    0    0    1     1     1    NA    NA
#>  [60,]    1    1    2    1    1    1    0    1    0     1     0     0     0
#>  [61,]    0    0    0    0    0    0    0    0   NA    NA    NA    NA    NA
#>  [62,]    2    0    0    0    2    0    0    1    2    NA    NA    NA    NA
#>  [63,]    0    1    0    2    3   NA   NA   NA   NA    NA    NA    NA    NA
#>  [64,]    1    1    1    0    1    0    1    0    0     0    NA    NA    NA
#>  [65,]    2    1    0    1    0    0    2   NA   NA    NA    NA    NA    NA
#>  [66,]    1    0    1    2    2    0    0   NA   NA    NA    NA    NA    NA
#>  [67,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [68,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [69,]    0    2    0    1    4    0    2    0    1     1     0     1     0
#>  [70,]    0    1    0    0    1    1    0    0    0    NA    NA    NA    NA
#>  [71,]    1    0    0    0    1    0    1    2    0     2     0     1    NA
#>  [72,]    1    2    0    1   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [73,]    0    0    0    2    0    0    1   NA   NA    NA    NA    NA    NA
#>  [74,]    0    0    0    0    0    0    0    0   NA    NA    NA    NA    NA
#>  [75,]    0    1    0    0    1    1   NA   NA   NA    NA    NA    NA    NA
#>  [76,]    0    0    0    0    0    0    0    0    0    NA    NA    NA    NA
#>  [77,]    1    0    3    0    3    0    1    2    0     1    NA    NA    NA
#>  [78,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [79,]    0    0    2    1    2    2    1    0    0     1     2    NA    NA
#>  [80,]    0    1    0    0    0    0    1    0    1     1     3     1     0
#>  [81,]    3    0    0    1    0    0    0   NA   NA    NA    NA    NA    NA
#>  [82,]    2    0    0    1    1    0    0    1    2     2     0     1     1
#>  [83,]    1    1   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [84,]    1    0    1    0    0    1    0    3    1    NA    NA    NA    NA
#>  [85,]    0    0    1    1    0    0    0    1    0     1     0     1     2
#>  [86,]    0    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [87,]    1    1    1    0    2    0    1   NA   NA    NA    NA    NA    NA
#>  [88,]    1    0    1   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [89,]    0    0    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [90,]    3    2    1   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [91,]    0    0    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [92,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [93,]    0    3    1    1    0    2    1    0    0     1     1     0     0
#>  [94,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [95,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [96,]    0    1    1    0    1    1   NA   NA   NA    NA    NA    NA    NA
#>  [97,]    0    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [98,]    2    1    0    0    1    0    0    0    0     1     1     0     2
#>  [99,]    0    1    0    2    1    0    2    0    2     1     2     1    NA
#> [100,]    2    0    0    1    1    1    0    0    1     1     2    NA    NA
#>        [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>   [1,]     0     0     0     0     0    NA    NA
#>   [2,]    NA    NA    NA    NA    NA    NA    NA
#>   [3,]    NA    NA    NA    NA    NA    NA    NA
#>   [4,]    NA    NA    NA    NA    NA    NA    NA
#>   [5,]    NA    NA    NA    NA    NA    NA    NA
#>   [6,]    NA    NA    NA    NA    NA    NA    NA
#>   [7,]    NA    NA    NA    NA    NA    NA    NA
#>   [8,]    NA    NA    NA    NA    NA    NA    NA
#>   [9,]    NA    NA    NA    NA    NA    NA    NA
#>  [10,]     1     1     1     1     1     1    NA
#>  [11,]    NA    NA    NA    NA    NA    NA    NA
#>  [12,]    NA    NA    NA    NA    NA    NA    NA
#>  [13,]     0     0     0     0     0     0    NA
#>  [14,]    NA    NA    NA    NA    NA    NA    NA
#>  [15,]     0    NA    NA    NA    NA    NA    NA
#>  [16,]    NA    NA    NA    NA    NA    NA    NA
#>  [17,]     0     1     1    NA    NA    NA    NA
#>  [18,]    NA    NA    NA    NA    NA    NA    NA
#>  [19,]     0     0     1     0     3     1    NA
#>  [20,]    NA    NA    NA    NA    NA    NA    NA
#>  [21,]     0     1     2     0     2     2     1
#>  [22,]    NA    NA    NA    NA    NA    NA    NA
#>  [23,]    NA    NA    NA    NA    NA    NA    NA
#>  [24,]    NA    NA    NA    NA    NA    NA    NA
#>  [25,]    NA    NA    NA    NA    NA    NA    NA
#>  [26,]     1    NA    NA    NA    NA    NA    NA
#>  [27,]    NA    NA    NA    NA    NA    NA    NA
#>  [28,]    NA    NA    NA    NA    NA    NA    NA
#>  [29,]    NA    NA    NA    NA    NA    NA    NA
#>  [30,]    NA    NA    NA    NA    NA    NA    NA
#>  [31,]    NA    NA    NA    NA    NA    NA    NA
#>  [32,]     2     0     1     1    NA    NA    NA
#>  [33,]    NA    NA    NA    NA    NA    NA    NA
#>  [34,]     0     0     0     0    NA    NA    NA
#>  [35,]    NA    NA    NA    NA    NA    NA    NA
#>  [36,]     0     0     0     0     0    NA    NA
#>  [37,]     0     2     2    NA    NA    NA    NA
#>  [38,]     0     0     0     0    NA    NA    NA
#>  [39,]    NA    NA    NA    NA    NA    NA    NA
#>  [40,]     0     0    NA    NA    NA    NA    NA
#>  [41,]     1     0     0     2     0     2     1
#>  [42,]    NA    NA    NA    NA    NA    NA    NA
#>  [43,]     2     0    NA    NA    NA    NA    NA
#>  [44,]    NA    NA    NA    NA    NA    NA    NA
#>  [45,]    NA    NA    NA    NA    NA    NA    NA
#>  [46,]    NA    NA    NA    NA    NA    NA    NA
#>  [47,]     0     0     0     0     0     0     0
#>  [48,]    NA    NA    NA    NA    NA    NA    NA
#>  [49,]    NA    NA    NA    NA    NA    NA    NA
#>  [50,]     0     0     0     2     0     2    NA
#>  [51,]    NA    NA    NA    NA    NA    NA    NA
#>  [52,]     1     2     2     0     2    NA    NA
#>  [53,]    NA    NA    NA    NA    NA    NA    NA
#>  [54,]    NA    NA    NA    NA    NA    NA    NA
#>  [55,]    NA    NA    NA    NA    NA    NA    NA
#>  [56,]     0     0     0     0     0     0    NA
#>  [57,]    NA    NA    NA    NA    NA    NA    NA
#>  [58,]    NA    NA    NA    NA    NA    NA    NA
#>  [59,]    NA    NA    NA    NA    NA    NA    NA
#>  [60,]     2     0     1    NA    NA    NA    NA
#>  [61,]    NA    NA    NA    NA    NA    NA    NA
#>  [62,]    NA    NA    NA    NA    NA    NA    NA
#>  [63,]    NA    NA    NA    NA    NA    NA    NA
#>  [64,]    NA    NA    NA    NA    NA    NA    NA
#>  [65,]    NA    NA    NA    NA    NA    NA    NA
#>  [66,]    NA    NA    NA    NA    NA    NA    NA
#>  [67,]    NA    NA    NA    NA    NA    NA    NA
#>  [68,]    NA    NA    NA    NA    NA    NA    NA
#>  [69,]     0     1     0     1    NA    NA    NA
#>  [70,]    NA    NA    NA    NA    NA    NA    NA
#>  [71,]    NA    NA    NA    NA    NA    NA    NA
#>  [72,]    NA    NA    NA    NA    NA    NA    NA
#>  [73,]    NA    NA    NA    NA    NA    NA    NA
#>  [74,]    NA    NA    NA    NA    NA    NA    NA
#>  [75,]    NA    NA    NA    NA    NA    NA    NA
#>  [76,]    NA    NA    NA    NA    NA    NA    NA
#>  [77,]    NA    NA    NA    NA    NA    NA    NA
#>  [78,]    NA    NA    NA    NA    NA    NA    NA
#>  [79,]    NA    NA    NA    NA    NA    NA    NA
#>  [80,]     0     1     1     1     1     0    NA
#>  [81,]    NA    NA    NA    NA    NA    NA    NA
#>  [82,]     0     1     1     1     0     1     1
#>  [83,]    NA    NA    NA    NA    NA    NA    NA
#>  [84,]    NA    NA    NA    NA    NA    NA    NA
#>  [85,]     1     1    NA    NA    NA    NA    NA
#>  [86,]    NA    NA    NA    NA    NA    NA    NA
#>  [87,]    NA    NA    NA    NA    NA    NA    NA
#>  [88,]    NA    NA    NA    NA    NA    NA    NA
#>  [89,]    NA    NA    NA    NA    NA    NA    NA
#>  [90,]    NA    NA    NA    NA    NA    NA    NA
#>  [91,]    NA    NA    NA    NA    NA    NA    NA
#>  [92,]    NA    NA    NA    NA    NA    NA    NA
#>  [93,]     1     1     2     1    NA    NA    NA
#>  [94,]    NA    NA    NA    NA    NA    NA    NA
#>  [95,]    NA    NA    NA    NA    NA    NA    NA
#>  [96,]    NA    NA    NA    NA    NA    NA    NA
#>  [97,]    NA    NA    NA    NA    NA    NA    NA
#>  [98,]     1    NA    NA    NA    NA    NA    NA
#>  [99,]    NA    NA    NA    NA    NA    NA    NA
#> [100,]    NA    NA    NA    NA    NA    NA    NA
#> 
#> , , 2
#> 
#>        [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>   [1,]    1    0    1    0    3    0    1    1    1     0     0     0     1
#>   [2,]    1    0    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>   [3,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>   [4,]    0    1    1    1   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>   [5,]    1    0    2    0    1    2    3    0   NA    NA    NA    NA    NA
#>   [6,]    1    0    2    0    0    0    1    0    0     0     0    NA    NA
#>   [7,]    0    1    1    0    1    1    0   NA   NA    NA    NA    NA    NA
#>   [8,]    2    1    1    0    1    2    1    1    0     3     2    NA    NA
#>   [9,]    1    2    0    1    1    0    0    2    1     0     0     1    NA
#>  [10,]    2    0    0    0    1    1    1    0    0     1     0     1     2
#>  [11,]    1    0    0    1    1    0    0    0    1     0     0    NA    NA
#>  [12,]    1    2    1    3    0    1    0    1    0     2    NA    NA    NA
#>  [13,]    1    1    0    0    1    1    1    0    2     1     0     0     0
#>  [14,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [15,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [16,]    0    0    1    1    2    1    0    2    1     1     0     0     2
#>  [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [18,]    0    0    1   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [19,]    0    0    0    1    0    0    0    0    2     1     1     1     0
#>  [20,]    0    1    1    0    1    1   NA   NA   NA    NA    NA    NA    NA
#>  [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [22,]    0    2   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [23,]    0    0    1    1    0    0    0    1    0     1     2     2     2
#>  [24,]    0    1    0    2    2    1   NA   NA   NA    NA    NA    NA    NA
#>  [25,]    0    0    0    0    0    0    0    0   NA    NA    NA    NA    NA
#>  [26,]    1    1    2    0    3    0    1    1    0     0     1     1     1
#>  [27,]    2    1    0    0    1   NA   NA   NA   NA    NA    NA    NA    NA
#>  [28,]    0    1    1    2    0    1    1    0    1     0    NA    NA    NA
#>  [29,]    1    1    0    2    1   NA   NA   NA   NA    NA    NA    NA    NA
#>  [30,]    0    2    3    0    2    0    0    0   NA    NA    NA    NA    NA
#>  [31,]    1    2    0    3    1    2    0    0    0    NA    NA    NA    NA
#>  [32,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [33,]    0    0    0    0    0    0    0    0    0     0     0    NA    NA
#>  [34,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [35,]    0    0    0    0    0    0    0    0    0    NA    NA    NA    NA
#>  [36,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [37,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [38,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [39,]    0    1    1    0    0    0    0    1    1     3     0     1     0
#>  [40,]    1    2    1    0    1    1    1    1    0     0     0     0     1
#>  [41,]    1    0    0    3    0    1    0    2    2     0     3     1     0
#>  [42,]    1    1    1    1   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [43,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [44,]    0    2    1    2   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [45,]    1    1    0    1    0    0    4    0    0     1     0     1     0
#>  [46,]    0    0    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [47,]    0    4    0    2    0    0    0    1    1     1     1     2     0
#>  [48,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [49,]    0    3    2    2    0    0    1    3   NA    NA    NA    NA    NA
#>  [50,]    1    2    0    0    1    0    2    0    0     0     1     0     1
#>  [51,]    0    0    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [52,]    0    0    0    0    1    2    1    2    0     0     1     0     1
#>  [53,]    0    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [54,]    1    2    1    0    2    0    0    0    1     1     2     1    NA
#>  [55,]    0    1    1    1    2    2    0    2    0     2    NA    NA    NA
#>  [56,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [57,]    0    1    0    3    1    0    0    1    0     1    NA    NA    NA
#>  [58,]    0    1    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [59,]    2    0    1    1    1    0    0    1    1     0     1    NA    NA
#>  [60,]    0    1    1    1    3    0    0    0    0     0     1     0     2
#>  [61,]    2    1    1    0    0    1    0    0   NA    NA    NA    NA    NA
#>  [62,]    0    0    0    0    0    0    0    0    0    NA    NA    NA    NA
#>  [63,]    0    0    0    0    0   NA   NA   NA   NA    NA    NA    NA    NA
#>  [64,]    0    0    0    0    0    0    0    0    0     0    NA    NA    NA
#>  [65,]    0    0    0    0    0    0    0   NA   NA    NA    NA    NA    NA
#>  [66,]    0    0    0    0    0    0    0   NA   NA    NA    NA    NA    NA
#>  [67,]    2    0    1    1    1    0    1    1    0     0     1     1     0
#>  [68,]    0    0    0    0    2    2    0    1    1     0     2     1     1
#>  [69,]    3    2    2    0    2    0    2    0    1     1     1     1     3
#>  [70,]    1    0    1    0    0    1    1    0    0    NA    NA    NA    NA
#>  [71,]    0    0    0    0    0    0    0    0    0     0     0     0    NA
#>  [72,]    1    0    0    1   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [73,]    4    0    2    0    0    1    1   NA   NA    NA    NA    NA    NA
#>  [74,]    0    0    0    1    2    1    2    0   NA    NA    NA    NA    NA
#>  [75,]    0    1    1    1    2    1   NA   NA   NA    NA    NA    NA    NA
#>  [76,]    0    1    1    2    0    3    1    1    0    NA    NA    NA    NA
#>  [77,]    0    2    1    3    0    1    1    1    0     1    NA    NA    NA
#>  [78,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [79,]    0    2    0    0    0    0    1    1    0     1     2    NA    NA
#>  [80,]    2    1    2    2    0    2    0    2    0     3     1     0     0
#>  [81,]    0    1    1    2    0    1    0   NA   NA    NA    NA    NA    NA
#>  [82,]    0    1    0    0    0    0    0    0    0     0     2     1     2
#>  [83,]    0    1   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [84,]    0    0    0    0    0    0    0    0    0    NA    NA    NA    NA
#>  [85,]    1    1    0    2    0    0    0    1    1     1     1     0     1
#>  [86,]    1    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [87,]    1    0    1    1    2    0    1   NA   NA    NA    NA    NA    NA
#>  [88,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [89,]    1    1    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [90,]    2    1    4   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [91,]    0    2    1    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [92,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [93,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [94,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [95,]    0    2    1   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [96,]    1    0    0    0    1    0   NA   NA   NA    NA    NA    NA    NA
#>  [97,]    1    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [98,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [99,]    0    0    0    0    0    0    1    1    0     0     0     1    NA
#> [100,]    1    1    0    0    1    0    1    1    1     0     1    NA    NA
#>        [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>   [1,]     1     0     0     1     1    NA    NA
#>   [2,]    NA    NA    NA    NA    NA    NA    NA
#>   [3,]    NA    NA    NA    NA    NA    NA    NA
#>   [4,]    NA    NA    NA    NA    NA    NA    NA
#>   [5,]    NA    NA    NA    NA    NA    NA    NA
#>   [6,]    NA    NA    NA    NA    NA    NA    NA
#>   [7,]    NA    NA    NA    NA    NA    NA    NA
#>   [8,]    NA    NA    NA    NA    NA    NA    NA
#>   [9,]    NA    NA    NA    NA    NA    NA    NA
#>  [10,]     0     0     3     0     0     1    NA
#>  [11,]    NA    NA    NA    NA    NA    NA    NA
#>  [12,]    NA    NA    NA    NA    NA    NA    NA
#>  [13,]     1     2     0     2     2     1    NA
#>  [14,]    NA    NA    NA    NA    NA    NA    NA
#>  [15,]     0    NA    NA    NA    NA    NA    NA
#>  [16,]    NA    NA    NA    NA    NA    NA    NA
#>  [17,]     0     0     0    NA    NA    NA    NA
#>  [18,]    NA    NA    NA    NA    NA    NA    NA
#>  [19,]     0     1     0     1     1     0    NA
#>  [20,]    NA    NA    NA    NA    NA    NA    NA
#>  [21,]     0     0     0     0     0     0     0
#>  [22,]    NA    NA    NA    NA    NA    NA    NA
#>  [23,]    NA    NA    NA    NA    NA    NA    NA
#>  [24,]    NA    NA    NA    NA    NA    NA    NA
#>  [25,]    NA    NA    NA    NA    NA    NA    NA
#>  [26,]     0    NA    NA    NA    NA    NA    NA
#>  [27,]    NA    NA    NA    NA    NA    NA    NA
#>  [28,]    NA    NA    NA    NA    NA    NA    NA
#>  [29,]    NA    NA    NA    NA    NA    NA    NA
#>  [30,]    NA    NA    NA    NA    NA    NA    NA
#>  [31,]    NA    NA    NA    NA    NA    NA    NA
#>  [32,]     0     0     0     0    NA    NA    NA
#>  [33,]    NA    NA    NA    NA    NA    NA    NA
#>  [34,]     0     0     0     0    NA    NA    NA
#>  [35,]    NA    NA    NA    NA    NA    NA    NA
#>  [36,]     0     0     0     0     0    NA    NA
#>  [37,]     0     0     0    NA    NA    NA    NA
#>  [38,]     0     0     0     0    NA    NA    NA
#>  [39,]    NA    NA    NA    NA    NA    NA    NA
#>  [40,]     4     1    NA    NA    NA    NA    NA
#>  [41,]     1     0     0     1     0     1     1
#>  [42,]    NA    NA    NA    NA    NA    NA    NA
#>  [43,]     0     0    NA    NA    NA    NA    NA
#>  [44,]    NA    NA    NA    NA    NA    NA    NA
#>  [45,]    NA    NA    NA    NA    NA    NA    NA
#>  [46,]    NA    NA    NA    NA    NA    NA    NA
#>  [47,]     0     1     1     0     1     2     0
#>  [48,]    NA    NA    NA    NA    NA    NA    NA
#>  [49,]    NA    NA    NA    NA    NA    NA    NA
#>  [50,]     0     1     1     1     0     0    NA
#>  [51,]    NA    NA    NA    NA    NA    NA    NA
#>  [52,]     2     1     1     0     0    NA    NA
#>  [53,]    NA    NA    NA    NA    NA    NA    NA
#>  [54,]    NA    NA    NA    NA    NA    NA    NA
#>  [55,]    NA    NA    NA    NA    NA    NA    NA
#>  [56,]     0     0     0     0     0     0    NA
#>  [57,]    NA    NA    NA    NA    NA    NA    NA
#>  [58,]    NA    NA    NA    NA    NA    NA    NA
#>  [59,]    NA    NA    NA    NA    NA    NA    NA
#>  [60,]     0     0     1    NA    NA    NA    NA
#>  [61,]    NA    NA    NA    NA    NA    NA    NA
#>  [62,]    NA    NA    NA    NA    NA    NA    NA
#>  [63,]    NA    NA    NA    NA    NA    NA    NA
#>  [64,]    NA    NA    NA    NA    NA    NA    NA
#>  [65,]    NA    NA    NA    NA    NA    NA    NA
#>  [66,]    NA    NA    NA    NA    NA    NA    NA
#>  [67,]    NA    NA    NA    NA    NA    NA    NA
#>  [68,]    NA    NA    NA    NA    NA    NA    NA
#>  [69,]     0     1     3     1    NA    NA    NA
#>  [70,]    NA    NA    NA    NA    NA    NA    NA
#>  [71,]    NA    NA    NA    NA    NA    NA    NA
#>  [72,]    NA    NA    NA    NA    NA    NA    NA
#>  [73,]    NA    NA    NA    NA    NA    NA    NA
#>  [74,]    NA    NA    NA    NA    NA    NA    NA
#>  [75,]    NA    NA    NA    NA    NA    NA    NA
#>  [76,]    NA    NA    NA    NA    NA    NA    NA
#>  [77,]    NA    NA    NA    NA    NA    NA    NA
#>  [78,]    NA    NA    NA    NA    NA    NA    NA
#>  [79,]    NA    NA    NA    NA    NA    NA    NA
#>  [80,]     0     1     1     0     0     0    NA
#>  [81,]    NA    NA    NA    NA    NA    NA    NA
#>  [82,]     2     0     0     1     0     1     0
#>  [83,]    NA    NA    NA    NA    NA    NA    NA
#>  [84,]    NA    NA    NA    NA    NA    NA    NA
#>  [85,]     1     0    NA    NA    NA    NA    NA
#>  [86,]    NA    NA    NA    NA    NA    NA    NA
#>  [87,]    NA    NA    NA    NA    NA    NA    NA
#>  [88,]    NA    NA    NA    NA    NA    NA    NA
#>  [89,]    NA    NA    NA    NA    NA    NA    NA
#>  [90,]    NA    NA    NA    NA    NA    NA    NA
#>  [91,]    NA    NA    NA    NA    NA    NA    NA
#>  [92,]    NA    NA    NA    NA    NA    NA    NA
#>  [93,]     0     0     0     0    NA    NA    NA
#>  [94,]    NA    NA    NA    NA    NA    NA    NA
#>  [95,]    NA    NA    NA    NA    NA    NA    NA
#>  [96,]    NA    NA    NA    NA    NA    NA    NA
#>  [97,]    NA    NA    NA    NA    NA    NA    NA
#>  [98,]     0    NA    NA    NA    NA    NA    NA
#>  [99,]    NA    NA    NA    NA    NA    NA    NA
#> [100,]    NA    NA    NA    NA    NA    NA    NA
#> 
#> , , 3
#> 
#>        [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>   [1,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>   [2,]    0    0    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>   [3,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>   [4,]    3    1    0    2   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>   [5,]    0    0    0    0    0    0    0    0   NA    NA    NA    NA    NA
#>   [6,]    1    1    0    1    0    0    0    2    1     0     1    NA    NA
#>   [7,]    0    0    0    0    0    0    0   NA   NA    NA    NA    NA    NA
#>   [8,]    0    0    0    0    0    0    0    0    0     0     0    NA    NA
#>   [9,]    0    0    0    0    0    0    0    0    0     0     0     0    NA
#>  [10,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [11,]    0    1    1    0    0    2    2    1    2     1     1    NA    NA
#>  [12,]    2    1    1    1    0    1    0    0    0     0    NA    NA    NA
#>  [13,]    0    0    1    2    1    1    1    1    0     0     0     0     0
#>  [14,]    1    0    0    0    1    1    1    1    1     2     0     1     1
#>  [15,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [16,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [17,]    1    2    0    1    2    1    3    0    2     0     1     0     3
#>  [18,]    2    0    1   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [19,]    0    0    2    1    0    0    0    0    0     1     1     0     1
#>  [20,]    0    1    1    0    1    0   NA   NA   NA    NA    NA    NA    NA
#>  [21,]    2    0    2    0    1    3    3    2    0     0     0     0     1
#>  [22,]    0    1   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [23,]    0    1    0    0    0    0    1    0    0     2     1     1     0
#>  [24,]    0    0    0    0    3    0   NA   NA   NA    NA    NA    NA    NA
#>  [25,]    2    0    1    1    0    2    0    0   NA    NA    NA    NA    NA
#>  [26,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [27,]    1    0    0    3    2   NA   NA   NA   NA    NA    NA    NA    NA
#>  [28,]    1    3    0    1    0    3    0    1    0     2    NA    NA    NA
#>  [29,]    0    1    0    0    0   NA   NA   NA   NA    NA    NA    NA    NA
#>  [30,]    1    0    0    0    1    2    2    0   NA    NA    NA    NA    NA
#>  [31,]    0    0    0    0    0    0    0    0    0    NA    NA    NA    NA
#>  [32,]    1    4    2    0    1    2    1    0    1     1     1     0     1
#>  [33,]    0    0    0    0    0    0    0    0    0     0     0    NA    NA
#>  [34,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [35,]    0    0    2    2    1    0    2    0    2    NA    NA    NA    NA
#>  [36,]    2    2    0    0    0    2    2    0    2     1     1     1     1
#>  [37,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [38,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [39,]    0    1    1    3    0    1    2    0    1     0     1     0     2
#>  [40,]    1    2    0    1    1    1    1    0    2     2     1     0     2
#>  [41,]    1    2    0    2    0    1    0    2    0     0     1     0     1
#>  [42,]    0    1    2    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [43,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [44,]    1    1    1    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [45,]    1    0    0    0    0    1    1    0    0     0     0     1     0
#>  [46,]    2    0    4    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [47,]    2    0    1    1    0    2    2    0    2     0     1     0     1
#>  [48,]    1    0    1   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [49,]    2    0    1    1    2    1    0    0   NA    NA    NA    NA    NA
#>  [50,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [51,]    0    2    2    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [52,]    0    0    1    1    0    3    1    0    0     2     0     0     0
#>  [53,]    2    2   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [54,]    1    2    1    3    0    0    3    1    0     1     1     0    NA
#>  [55,]    1    1    1    1    0    0    1    2    4     0    NA    NA    NA
#>  [56,]    2    1    2    0    0    0    0    0    2     0     1     2     1
#>  [57,]    0    0    0    0    0    0    0    0    0     0    NA    NA    NA
#>  [58,]    0    1    1    2   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [59,]    0    0    1    0    0    0    1    1    0     1     0    NA    NA
#>  [60,]    2    0    2    0    0    2    3    2    0     1     0     0     0
#>  [61,]    0    0    0    0    0    0    0    0   NA    NA    NA    NA    NA
#>  [62,]    0    0    0    0    0    0    0    0    0    NA    NA    NA    NA
#>  [63,]    0    0    0    0    0   NA   NA   NA   NA    NA    NA    NA    NA
#>  [64,]    0    0    0    0    0    0    0    0    0     0    NA    NA    NA
#>  [65,]    0    0    1    2    1    0    1   NA   NA    NA    NA    NA    NA
#>  [66,]    0    0    0    0    0    0    0   NA   NA    NA    NA    NA    NA
#>  [67,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [68,]    1    0    1    1    0    2    0    2    0     0     0     0     2
#>  [69,]    1    1    1    1    2    0    0    1    0     0     0     2     1
#>  [70,]    0    1    1    2    1    2    1    2    0    NA    NA    NA    NA
#>  [71,]    1    1    1    2    2    0    1    0    0     1     0     0    NA
#>  [72,]    0    0    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [73,]    0    0    0    2    0    0    6   NA   NA    NA    NA    NA    NA
#>  [74,]    1    2    1    2    1    1    1    1   NA    NA    NA    NA    NA
#>  [75,]    1    0    0    0    0    3   NA   NA   NA    NA    NA    NA    NA
#>  [76,]    0    0    1    0    0    1    0    2    1    NA    NA    NA    NA
#>  [77,]    1    1    0    1    1    3    0    2    0     1    NA    NA    NA
#>  [78,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [79,]    0    0    0    0    0    0    0    0    0     0     0    NA    NA
#>  [80,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [81,]    2    2    2    0    0    0    1   NA   NA    NA    NA    NA    NA
#>  [82,]    0    2    3    1    4    0    0    0    1     0     2     1     1
#>  [83,]    1    1   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [84,]    0    0    0    0    0    0    0    2    0    NA    NA    NA    NA
#>  [85,]    0    3    0    1    0    1    0    0    2     0     0     0     1
#>  [86,]    1    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [87,]    0    0    0    0    0    0    0   NA   NA    NA    NA    NA    NA
#>  [88,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [89,]    2    1    2    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [90,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [91,]    1    0    1    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [92,]    3    0    1    0    1    0    1    1    0     0     1     0     0
#>  [93,]    0    0    0    1    0    0    2    1    0     0     1     0     0
#>  [94,]    1    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [95,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [96,]    0    0    0    0    0    0   NA   NA   NA    NA    NA    NA    NA
#>  [97,]    0    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [98,]    1    0    1    0    0    3    0    1    0     1     1     3     1
#>  [99,]    0    0    1    0    1    2    0    0    1     4     1     0    NA
#> [100,]    1    3    0    1    1    1    2    0    1     1     1    NA    NA
#>        [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>   [1,]     0     0     0     0     0    NA    NA
#>   [2,]    NA    NA    NA    NA    NA    NA    NA
#>   [3,]    NA    NA    NA    NA    NA    NA    NA
#>   [4,]    NA    NA    NA    NA    NA    NA    NA
#>   [5,]    NA    NA    NA    NA    NA    NA    NA
#>   [6,]    NA    NA    NA    NA    NA    NA    NA
#>   [7,]    NA    NA    NA    NA    NA    NA    NA
#>   [8,]    NA    NA    NA    NA    NA    NA    NA
#>   [9,]    NA    NA    NA    NA    NA    NA    NA
#>  [10,]     0     0     0     0     0     0    NA
#>  [11,]    NA    NA    NA    NA    NA    NA    NA
#>  [12,]    NA    NA    NA    NA    NA    NA    NA
#>  [13,]     0     0     1     2     2     0    NA
#>  [14,]    NA    NA    NA    NA    NA    NA    NA
#>  [15,]     0    NA    NA    NA    NA    NA    NA
#>  [16,]    NA    NA    NA    NA    NA    NA    NA
#>  [17,]     1     0     0    NA    NA    NA    NA
#>  [18,]    NA    NA    NA    NA    NA    NA    NA
#>  [19,]     2     0     1     1     0     1    NA
#>  [20,]    NA    NA    NA    NA    NA    NA    NA
#>  [21,]     0     1     0     0     0     1     0
#>  [22,]    NA    NA    NA    NA    NA    NA    NA
#>  [23,]    NA    NA    NA    NA    NA    NA    NA
#>  [24,]    NA    NA    NA    NA    NA    NA    NA
#>  [25,]    NA    NA    NA    NA    NA    NA    NA
#>  [26,]     0    NA    NA    NA    NA    NA    NA
#>  [27,]    NA    NA    NA    NA    NA    NA    NA
#>  [28,]    NA    NA    NA    NA    NA    NA    NA
#>  [29,]    NA    NA    NA    NA    NA    NA    NA
#>  [30,]    NA    NA    NA    NA    NA    NA    NA
#>  [31,]    NA    NA    NA    NA    NA    NA    NA
#>  [32,]     0     2     1     2    NA    NA    NA
#>  [33,]    NA    NA    NA    NA    NA    NA    NA
#>  [34,]     0     0     0     0    NA    NA    NA
#>  [35,]    NA    NA    NA    NA    NA    NA    NA
#>  [36,]     3     0     1     1     0    NA    NA
#>  [37,]     0     0     0    NA    NA    NA    NA
#>  [38,]     0     0     0     0    NA    NA    NA
#>  [39,]    NA    NA    NA    NA    NA    NA    NA
#>  [40,]     1     2    NA    NA    NA    NA    NA
#>  [41,]     0     1     0     1     0     0     0
#>  [42,]    NA    NA    NA    NA    NA    NA    NA
#>  [43,]     0     0    NA    NA    NA    NA    NA
#>  [44,]    NA    NA    NA    NA    NA    NA    NA
#>  [45,]    NA    NA    NA    NA    NA    NA    NA
#>  [46,]    NA    NA    NA    NA    NA    NA    NA
#>  [47,]     0     2     0     0     1     0     1
#>  [48,]    NA    NA    NA    NA    NA    NA    NA
#>  [49,]    NA    NA    NA    NA    NA    NA    NA
#>  [50,]     0     0     0     0     0     0    NA
#>  [51,]    NA    NA    NA    NA    NA    NA    NA
#>  [52,]     3     0     0     0     2    NA    NA
#>  [53,]    NA    NA    NA    NA    NA    NA    NA
#>  [54,]    NA    NA    NA    NA    NA    NA    NA
#>  [55,]    NA    NA    NA    NA    NA    NA    NA
#>  [56,]     2     0     1     2     1     0    NA
#>  [57,]    NA    NA    NA    NA    NA    NA    NA
#>  [58,]    NA    NA    NA    NA    NA    NA    NA
#>  [59,]    NA    NA    NA    NA    NA    NA    NA
#>  [60,]     0     1     3    NA    NA    NA    NA
#>  [61,]    NA    NA    NA    NA    NA    NA    NA
#>  [62,]    NA    NA    NA    NA    NA    NA    NA
#>  [63,]    NA    NA    NA    NA    NA    NA    NA
#>  [64,]    NA    NA    NA    NA    NA    NA    NA
#>  [65,]    NA    NA    NA    NA    NA    NA    NA
#>  [66,]    NA    NA    NA    NA    NA    NA    NA
#>  [67,]    NA    NA    NA    NA    NA    NA    NA
#>  [68,]    NA    NA    NA    NA    NA    NA    NA
#>  [69,]     0     0     2     1    NA    NA    NA
#>  [70,]    NA    NA    NA    NA    NA    NA    NA
#>  [71,]    NA    NA    NA    NA    NA    NA    NA
#>  [72,]    NA    NA    NA    NA    NA    NA    NA
#>  [73,]    NA    NA    NA    NA    NA    NA    NA
#>  [74,]    NA    NA    NA    NA    NA    NA    NA
#>  [75,]    NA    NA    NA    NA    NA    NA    NA
#>  [76,]    NA    NA    NA    NA    NA    NA    NA
#>  [77,]    NA    NA    NA    NA    NA    NA    NA
#>  [78,]    NA    NA    NA    NA    NA    NA    NA
#>  [79,]    NA    NA    NA    NA    NA    NA    NA
#>  [80,]     0     0     0     0     0     0    NA
#>  [81,]    NA    NA    NA    NA    NA    NA    NA
#>  [82,]     0     2     0     1     2     0     0
#>  [83,]    NA    NA    NA    NA    NA    NA    NA
#>  [84,]    NA    NA    NA    NA    NA    NA    NA
#>  [85,]     3     0    NA    NA    NA    NA    NA
#>  [86,]    NA    NA    NA    NA    NA    NA    NA
#>  [87,]    NA    NA    NA    NA    NA    NA    NA
#>  [88,]    NA    NA    NA    NA    NA    NA    NA
#>  [89,]    NA    NA    NA    NA    NA    NA    NA
#>  [90,]    NA    NA    NA    NA    NA    NA    NA
#>  [91,]    NA    NA    NA    NA    NA    NA    NA
#>  [92,]    NA    NA    NA    NA    NA    NA    NA
#>  [93,]     0     2     1     0    NA    NA    NA
#>  [94,]    NA    NA    NA    NA    NA    NA    NA
#>  [95,]    NA    NA    NA    NA    NA    NA    NA
#>  [96,]    NA    NA    NA    NA    NA    NA    NA
#>  [97,]    NA    NA    NA    NA    NA    NA    NA
#>  [98,]     1    NA    NA    NA    NA    NA    NA
#>  [99,]    NA    NA    NA    NA    NA    NA    NA
#> [100,]    NA    NA    NA    NA    NA    NA    NA
#> 
#> , , 4
#> 
#>        [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>   [1,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>   [2,]    2    1    0    2   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>   [3,]    0    3    1    1    0    1    0    2    0     0     1     0     2
#>   [4,]    0    0    1    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>   [5,]    0    1    2    0    1    0    3    0   NA    NA    NA    NA    NA
#>   [6,]    0    0    0    0    0    0    0    0    0     0     0    NA    NA
#>   [7,]    1    0    1    1    0    3    0   NA   NA    NA    NA    NA    NA
#>   [8,]    1    2    0    0    2    2    2    0    2     3     0    NA    NA
#>   [9,]    0    0    0    1    1    2    2    1    0     3     0     1    NA
#>  [10,]    0    0    0    0    0    1    2    1    1     1     1     0     1
#>  [11,]    0    0    0    0    0    0    0    0    0     0     0    NA    NA
#>  [12,]    0    0    0    0    0    1    0    0    0     1    NA    NA    NA
#>  [13,]    3    2    2    1    1    1    1    1    0     0     0     0     2
#>  [14,]    2    2    0    1    0    1    1    1    1     2     0     0     1
#>  [15,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [16,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [17,]    0    0    1    2    0    1    1    0    2     0     2     2     0
#>  [18,]    0    1    4   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [19,]    0    1    0    1    2    1    0    2    0     1     1     1     1
#>  [20,]    0    0    0    0    0    0   NA   NA   NA    NA    NA    NA    NA
#>  [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [22,]    1    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [23,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [24,]    0    1    0    0    1    3   NA   NA   NA    NA    NA    NA    NA
#>  [25,]    0    0    0    0    0    0    0    0   NA    NA    NA    NA    NA
#>  [26,]    1    3    1    0    1    0    2    0    0     0     2     0     1
#>  [27,]    1    1    0    0    1   NA   NA   NA   NA    NA    NA    NA    NA
#>  [28,]    0    0    1    0    2    0    1    0    1     0    NA    NA    NA
#>  [29,]    0    2    0    1    0   NA   NA   NA   NA    NA    NA    NA    NA
#>  [30,]    2    0    2    1    2    0    0    0   NA    NA    NA    NA    NA
#>  [31,]    0    0    0    0    0    0    0    0    0    NA    NA    NA    NA
#>  [32,]    4    1    1    1    1    2    0    1    0     1     2     0     0
#>  [33,]    0    0    0    0    0    0    0    0    0     0     0    NA    NA
#>  [34,]    1    1    1    0    2    0    1    0    0     1     0     2     1
#>  [35,]    0    0    0    0    0    0    0    0    0    NA    NA    NA    NA
#>  [36,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [37,]    0    0    0    1    0    1    2    2    0     2     1     1     0
#>  [38,]    0    0    2    0    0    1    1    2    2     3     0     2     2
#>  [39,]    1    0    1    1    0    0    1    1    0     1     0     1     0
#>  [40,]    0    1    0    0    1    1    1    0    1     2     2     0     1
#>  [41,]    0    1    0    0    0    3    1    0    1     1     1     2     1
#>  [42,]    2    0    1    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [43,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [44,]    0    2    0    2   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [45,]    0    1    1    1    0    0    1    0    0     0     0     1     1
#>  [46,]    1    1    2    2   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [47,]    0    1    0    0    0    0    2    2    3     0     0     0     0
#>  [48,]    1    2    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [49,]    3    0    1    1    0    2    1    1   NA    NA    NA    NA    NA
#>  [50,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [51,]    0    2    0    1   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [52,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [53,]    0    1   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [54,]    0    1    0    0    0    1    1    0    1     1     0     0    NA
#>  [55,]    0    2    0    0    2    0    2    4    1     0    NA    NA    NA
#>  [56,]    0    3    0    3    2    0    0    1    0     1     0     2     1
#>  [57,]    0    0    0    0    0    0    0    0    0     0    NA    NA    NA
#>  [58,]    0    0    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [59,]    0    1    1    2    0    0    1    0    0     1     0    NA    NA
#>  [60,]    0    1    3    1    2    2    1    1    1     0     0     1     1
#>  [61,]    0    1    2    0    0    0    1    0   NA    NA    NA    NA    NA
#>  [62,]    0    0    0    0    0    0    0    0    0    NA    NA    NA    NA
#>  [63,]    1    1    0    1    1   NA   NA   NA   NA    NA    NA    NA    NA
#>  [64,]    0    0    0    0    0    0    0    0    0     0    NA    NA    NA
#>  [65,]    1    1    1    3    1    1    0   NA   NA    NA    NA    NA    NA
#>  [66,]    0    0    0    1    2    0    2   NA   NA    NA    NA    NA    NA
#>  [67,]    0    1    1    1    1    0    2    1    0     0     0     3     0
#>  [68,]    0    1    1    2    1    1    0    2    2     0     1     1     0
#>  [69,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [70,]    1    0    1    1    0    1    1    0    1    NA    NA    NA    NA
#>  [71,]    0    0    0    0    0    0    0    0    0     0     0     0    NA
#>  [72,]    1    2    0    2   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [73,]    0    0    0    0    2    0    0   NA   NA    NA    NA    NA    NA
#>  [74,]    0    0    1    1    0    0    0    0   NA    NA    NA    NA    NA
#>  [75,]    0    0    1    1    2    1   NA   NA   NA    NA    NA    NA    NA
#>  [76,]    1    1    0    0    0    1    1    0    1    NA    NA    NA    NA
#>  [77,]    0    0    0    0    0    0    0    0    0     0    NA    NA    NA
#>  [78,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [79,]    0    0    2    1    2    1    1    1    1     2     3    NA    NA
#>  [80,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [81,]    1    1    2    0    2    0    0   NA   NA    NA    NA    NA    NA
#>  [82,]    0    1    0    0    1    3    0    0    0     1     1     3     0
#>  [83,]    0    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [84,]    0    0    0    0    0    0    0    0    0    NA    NA    NA    NA
#>  [85,]    2    3    2    0    0    0    1    0    2     1     0     1     2
#>  [86,]    1    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [87,]    1    2    3    0    0    0    0   NA   NA    NA    NA    NA    NA
#>  [88,]    0    2    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [89,]    0    0    3    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [90,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [91,]    0    3    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [92,]    1    1    1    0    1    0    1    0    0     0     0     1     0
#>  [93,]    3    0    2    0    0    0    0    0    0     0     1     0     1
#>  [94,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [95,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [96,]    0    0    1    0    1    3   NA   NA   NA    NA    NA    NA    NA
#>  [97,]    1    1   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [98,]    0    1    2    2    2    1    3    0    0     2     2     0     1
#>  [99,]    0    1    1    1    0    0    0    2    3     0     2     0    NA
#> [100,]    0    4    2    1    1    0    0    0    1     2     0    NA    NA
#>        [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>   [1,]     0     0     0     0     0    NA    NA
#>   [2,]    NA    NA    NA    NA    NA    NA    NA
#>   [3,]    NA    NA    NA    NA    NA    NA    NA
#>   [4,]    NA    NA    NA    NA    NA    NA    NA
#>   [5,]    NA    NA    NA    NA    NA    NA    NA
#>   [6,]    NA    NA    NA    NA    NA    NA    NA
#>   [7,]    NA    NA    NA    NA    NA    NA    NA
#>   [8,]    NA    NA    NA    NA    NA    NA    NA
#>   [9,]    NA    NA    NA    NA    NA    NA    NA
#>  [10,]     3     1     1     1     1     1    NA
#>  [11,]    NA    NA    NA    NA    NA    NA    NA
#>  [12,]    NA    NA    NA    NA    NA    NA    NA
#>  [13,]     0     2     0     3     2     0    NA
#>  [14,]    NA    NA    NA    NA    NA    NA    NA
#>  [15,]     0    NA    NA    NA    NA    NA    NA
#>  [16,]    NA    NA    NA    NA    NA    NA    NA
#>  [17,]     1     0     1    NA    NA    NA    NA
#>  [18,]    NA    NA    NA    NA    NA    NA    NA
#>  [19,]     2     0     0     1     1     0    NA
#>  [20,]    NA    NA    NA    NA    NA    NA    NA
#>  [21,]     0     0     0     0     0     0     0
#>  [22,]    NA    NA    NA    NA    NA    NA    NA
#>  [23,]    NA    NA    NA    NA    NA    NA    NA
#>  [24,]    NA    NA    NA    NA    NA    NA    NA
#>  [25,]    NA    NA    NA    NA    NA    NA    NA
#>  [26,]     0    NA    NA    NA    NA    NA    NA
#>  [27,]    NA    NA    NA    NA    NA    NA    NA
#>  [28,]    NA    NA    NA    NA    NA    NA    NA
#>  [29,]    NA    NA    NA    NA    NA    NA    NA
#>  [30,]    NA    NA    NA    NA    NA    NA    NA
#>  [31,]    NA    NA    NA    NA    NA    NA    NA
#>  [32,]     1     2     1     0    NA    NA    NA
#>  [33,]    NA    NA    NA    NA    NA    NA    NA
#>  [34,]     1     1     1     1    NA    NA    NA
#>  [35,]    NA    NA    NA    NA    NA    NA    NA
#>  [36,]     0     0     0     0     0    NA    NA
#>  [37,]     0     2     0    NA    NA    NA    NA
#>  [38,]     2     1     0     1    NA    NA    NA
#>  [39,]    NA    NA    NA    NA    NA    NA    NA
#>  [40,]     0     1    NA    NA    NA    NA    NA
#>  [41,]     1     0     1     0     0     1     1
#>  [42,]    NA    NA    NA    NA    NA    NA    NA
#>  [43,]     0     0    NA    NA    NA    NA    NA
#>  [44,]    NA    NA    NA    NA    NA    NA    NA
#>  [45,]    NA    NA    NA    NA    NA    NA    NA
#>  [46,]    NA    NA    NA    NA    NA    NA    NA
#>  [47,]     0     0     0     2     2     2     0
#>  [48,]    NA    NA    NA    NA    NA    NA    NA
#>  [49,]    NA    NA    NA    NA    NA    NA    NA
#>  [50,]     0     0     0     0     0     0    NA
#>  [51,]    NA    NA    NA    NA    NA    NA    NA
#>  [52,]     0     0     0     0     0    NA    NA
#>  [53,]    NA    NA    NA    NA    NA    NA    NA
#>  [54,]    NA    NA    NA    NA    NA    NA    NA
#>  [55,]    NA    NA    NA    NA    NA    NA    NA
#>  [56,]     2     1     2     1     1     2    NA
#>  [57,]    NA    NA    NA    NA    NA    NA    NA
#>  [58,]    NA    NA    NA    NA    NA    NA    NA
#>  [59,]    NA    NA    NA    NA    NA    NA    NA
#>  [60,]     0     1     0    NA    NA    NA    NA
#>  [61,]    NA    NA    NA    NA    NA    NA    NA
#>  [62,]    NA    NA    NA    NA    NA    NA    NA
#>  [63,]    NA    NA    NA    NA    NA    NA    NA
#>  [64,]    NA    NA    NA    NA    NA    NA    NA
#>  [65,]    NA    NA    NA    NA    NA    NA    NA
#>  [66,]    NA    NA    NA    NA    NA    NA    NA
#>  [67,]    NA    NA    NA    NA    NA    NA    NA
#>  [68,]    NA    NA    NA    NA    NA    NA    NA
#>  [69,]     0     0     0     0    NA    NA    NA
#>  [70,]    NA    NA    NA    NA    NA    NA    NA
#>  [71,]    NA    NA    NA    NA    NA    NA    NA
#>  [72,]    NA    NA    NA    NA    NA    NA    NA
#>  [73,]    NA    NA    NA    NA    NA    NA    NA
#>  [74,]    NA    NA    NA    NA    NA    NA    NA
#>  [75,]    NA    NA    NA    NA    NA    NA    NA
#>  [76,]    NA    NA    NA    NA    NA    NA    NA
#>  [77,]    NA    NA    NA    NA    NA    NA    NA
#>  [78,]    NA    NA    NA    NA    NA    NA    NA
#>  [79,]    NA    NA    NA    NA    NA    NA    NA
#>  [80,]     0     0     0     0     0     0    NA
#>  [81,]    NA    NA    NA    NA    NA    NA    NA
#>  [82,]     1     2     0     1     2     0     0
#>  [83,]    NA    NA    NA    NA    NA    NA    NA
#>  [84,]    NA    NA    NA    NA    NA    NA    NA
#>  [85,]     1     1    NA    NA    NA    NA    NA
#>  [86,]    NA    NA    NA    NA    NA    NA    NA
#>  [87,]    NA    NA    NA    NA    NA    NA    NA
#>  [88,]    NA    NA    NA    NA    NA    NA    NA
#>  [89,]    NA    NA    NA    NA    NA    NA    NA
#>  [90,]    NA    NA    NA    NA    NA    NA    NA
#>  [91,]    NA    NA    NA    NA    NA    NA    NA
#>  [92,]    NA    NA    NA    NA    NA    NA    NA
#>  [93,]     2     0     2     0    NA    NA    NA
#>  [94,]    NA    NA    NA    NA    NA    NA    NA
#>  [95,]    NA    NA    NA    NA    NA    NA    NA
#>  [96,]    NA    NA    NA    NA    NA    NA    NA
#>  [97,]    NA    NA    NA    NA    NA    NA    NA
#>  [98,]     1    NA    NA    NA    NA    NA    NA
#>  [99,]    NA    NA    NA    NA    NA    NA    NA
#> [100,]    NA    NA    NA    NA    NA    NA    NA
#>