Fit the single species, single season site occupancy model.

This function fits the single species, single season site occupancy model first developed by MacKenzie et al. (2002).

occ_mod(
  occupancy,
  detection,
  data,
  niter = 1000,
  nchains = 3,
  seed = NULL,
  save_model = FALSE,
  model_name = paste0("occ_model_", Sys.Date())
)

Arguments

occupancy
model declaration for the occupancy portion of the model using standard linear model syntax

detection
model declaration for the detection portion of the model using standard linear model syntax

data
data from which to create the detection and covariate matrices. Each row should represent a single visit to a site. See details and examples for proper formatting

niter
number of MCMC iterations

seed
optional seed

save_model
logical; should a text file containing the model be exported?

model_name
character string defining the name of the text file describing the model if save_model = TRUE

Value

an object of class list containing the following:

samples object of class list of length nchains, each containing a matrix posterior samples

loglik object of class list of length codenchains, each containing a matrix of samples of the log posterior likelihood

Details

This function fits the single season, single species site occupancy model using the logit link function. The data should contain columns named site, visit, and y. See examples.

Examples

# simulate data
sim <- sim_occ(M = 250, max_j = 10, seed = 01012021, rand_visits = FALSE)
data <- sim$data

# note structure of data frame
names(data)

#> [1] "site"     "visit"    "y"        "psi_cov1" "p_cov1"  

head(data)

#>   site visit y   psi_cov1    p_cov1
#> 1    1     1 1 0.07477433 1.0225061
#> 2    1     2 0 0.07477433 1.0858551
#> 3    1     3 0 0.07477433 0.3505194
#> 4    1     4 0 0.07477433 0.3793353
#> 5    1     5 0 0.07477433 0.6347882
#> 6    1     6 1 0.07477433 1.3194706

# fit model
ex <- occ_mod(occupancy = ~ psi_cov1, detection = ~ p_cov1, data = data,
              niter = 2000)

#> Writing NIMBLE model to current working directory:

#> # priors
#> for(i in 1:2){beta_psi[i] ~ dnorm(0, 1/2)}
#> for(i in 1:2){beta_p[i] ~ dnorm(0, 1/2)}
#> 
#> # likelihood 
#> # loop through sites 
#> for(i in 1:M){ 
#>  # calculate psi 
#>  logit(psi[i]) <- beta_psi[1] +
#>   beta_psi[2] * psi_cov1[i]
#> 
#>  # latent occupancy state 
#>  z[i] ~ dbern(psi[i]) 
#> 
#>  # loop through visits 
#>  for(j in 1:nvisits[i]){
#>   # calculate p
#>   logit(p[i,j]) <- beta_p[1] +
#>    beta_p[2] * p_cov1[i,j]
#> 
#>   # response model
#>   Y[i,j] ~ dbern(z[i] * p[i,j])
#>  }
#> }

#> 
#> Building R model

#> defining model...

#> Detected Y as data within 'constants'.

#> Adding Y as data for building model.

#> building model...

#> setting data and initial values...

#> running calculate on model (any error reports that follow may simply reflect missing values in model variables) ... 

#> 

#> checking model sizes and dimensions...

#> 

#> model building finished.

#> 
#> Compiling R model to C

#> compiling... this may take a minute. Use 'showCompilerOutput = TRUE' to see C++ compilation details.

#> compilation finished.

#> 
#> Building R MCMC object

#> 
#> Compiling R MCMC object to C

#> compiling... this may take a minute. Use 'showCompilerOutput = TRUE' to see C++ compilation details.

#> compilation finished.

#> 
#> Running MCMC

#> running chain 1...

#> |-------------|-------------|-------------|-------------|
#> |-------------------------------------------------------|

#> running chain 2...

#> |-------------|-------------|-------------|-------------|
#> |-------------------------------------------------------|

#> running chain 3...

#> |-------------|-------------|-------------|-------------|
#> |-------------------------------------------------------|

#> 
#> Computing posterior likelihood

# results
str(ex)

#> List of 2
#>  $ samples:List of 3
#>   ..$ chain1: num [1:2000, 1:254] 0.621 0.621 0.99 0.99 0.99 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:254] "beta_p[1]" "beta_p[2]" "beta_psi[1]" "beta_psi[2]" ...
#>   ..$ chain2: num [1:2000, 1:254] -1.1022 -0.8708 -0.0517 -0.0517 -0.5245 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:254] "beta_p[1]" "beta_p[2]" "beta_psi[1]" "beta_psi[2]" ...
#>   ..$ chain3: num [1:2000, 1:254] 1.18 1.18 1.18 1.18 1.18 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:254] "beta_p[1]" "beta_p[2]" "beta_psi[1]" "beta_psi[2]" ...
#>  $ loglik :List of 3
#>   ..$ chain1: num [1:2000, 1:250] -9.25 -9.25 -8.72 -8.16 -8.12 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:250] "site1" "site2" "site3" "site4" ...
#>   ..$ chain2: num [1:2000, 1:250] -7.6 -6.38 -7.47 -7.47 -6.55 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:250] "site1" "site2" "site3" "site4" ...
#>   ..$ chain3: num [1:2000, 1:250] -8.5 -8.5 -8.7 -8.74 -8.74 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:250] "site1" "site2" "site3" "site4" ...
#>  - attr(*, "class")= chr [1:2] "occ_mod" "list"
#>  - attr(*, "code")= chr "# priors\nfor(i in 1:2){beta_psi[i] ~ dnorm(0, 1/2)}\nfor(i in 1:2){beta_p[i] ~ dnorm(0, 1/2)}\n\n# likelihood "| __truncated__
#>  - attr(*, "mcmcinfo")= chr [1:8] "===== Monitors =====" "thin = 1: z, beta_psi, beta_p" "===== Samplers =====" "RW sampler (4)" ...

ex

#> Single species, single season site occupancy model fit using NIMBLE.
#> 
#> The following model was fit:
#> # priors
#> for(i in 1:2){beta_psi[i] ~ dnorm(0, 1/2)}
#> for(i in 1:2){beta_p[i] ~ dnorm(0, 1/2)}
#> 
#> # likelihood 
#> # loop through sites 
#> for(i in 1:M){ 
#>  # calculate psi 
#>  logit(psi[i]) <- beta_psi[1] +
#>   beta_psi[2] * psi_cov1[i]
#> 
#>  # latent occupancy state 
#>  z[i] ~ dbern(psi[i]) 
#> 
#>  # loop through visits 
#>  for(j in 1:nvisits[i]){
#>   # calculate p
#>   logit(p[i,j]) <- beta_p[1] +
#>    beta_p[2] * p_cov1[i,j]
#> 
#>   # response model
#>   Y[i,j] ~ dbern(z[i] * p[i,j])
#>  }
#> }
#> 
#> The following MCMC algorithm was used:
#> ===== Monitors =====
#>  thin = 1: z, beta_psi, beta_p
#>  ===== Samplers =====
#>  RW sampler (4)
#>    - beta_psi[]  (2 elements)
#>    - beta_p[]  (2 elements)
#>  binary sampler (250)
#>    - z[]  (250 elements)

# lppd <- sum(log(colMeans(exp(loglik))))
# pwaic1 <- 2 * sum(log(colMeans(exp(loglik))) - colMeans(loglik))
# pwaic2 <- sum(apply(loglik, 2, var))
# waic1 <- -2 * (lppd - pwaic1)
# waic2 <- -2 * (lppd - pwaic2)