Simulate data from the single species, single season site occupancy model.

This function simulates data from the single species, single season occupancy model first developed by MacKenzie et al. (2002).

sim_occ(
  M = 20,
  max_j = 10,
  beta_psi = c(0, 1),
  beta_p = c(0, 1),
  seed = NULL,
  rand_visits = TRUE
)

Arguments

M
number of sites

max_j
maximum number of visits to each site. If rand_visits = FALSE, this value is the number of visits to each site. If rand_visits = TRUE, each site is visited a random (sample(2:max_j, size = 1)) number of times.

beta_psi
vector of regression coefficients used to generate psi

beta_p
vector of regression coefficients used to generate p

seed
optional seed for reproducibility

rand_visits
logical; should each site be visited a random number of times? See details.

Value

object of class list containing the following elements:

beta_psi vector of regression coefficients used to generate psi

beta_p vector of regression coefficients used to generate p

psi_cov matrix of site level covariates

p_cov array of detection level covariates; each slice represents a single covariate

psi vector of derived site level occupancy probabilities

p matrix of derived visit level detection probabilities

z vector of latent occupancy states for each site

Y matrix of observed Bernoulli responses

n_visits vector of number of visits to each site

data a data frame containing all information necessary to fit the model

Details

This function simulates data from the vanilla single season, single species occupancy model using the logit link function. If rand_visits = TRUE, each site is visited a random number of times between two and max_j. Covariates are drawn from the uniform(0, 1.5) distribution so that the effect of the direction of each regression coefficient is intuitive. Note that if covariates are not desired, beta_psi and beta_p can be set to intercepts that generate the desired derived probabilities.

Examples

sim <- sim_occ(M = 100, max_j = 20, beta_psi = c(0, 1), beta_p = c(0, 1, 1), rand_visits = TRUE)
sim$Y

#>        [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>   [1,]    0    1    1    1    1    1    1    1    1     1     1     1     1
#>   [2,]    1    0    1    1    0    1    1    1    1     1     1     1     1
#>   [3,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>   [4,]    1    1    0    1    1    0    1    1    1     1     0    NA    NA
#>   [5,]    0    1    1    1    1    1    1    1    0     1     1     1     0
#>   [6,]    1    1    1    1    0    1    1    1    1     1     1     0     1
#>   [7,]    1    0    0    1    1    0    1    1    1     1     1     1     1
#>   [8,]    1    0    1    1    1    1    1    1    1     1     1     0     0
#>   [9,]    1    1    1    1    1    1    1    1   NA    NA    NA    NA    NA
#>  [10,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [11,]    1    1    1    1    1    1    1   NA   NA    NA    NA    NA    NA
#>  [12,]    1    1    0    1    1    1    1    1   NA    NA    NA    NA    NA
#>  [13,]    1    1    1    1    1    1    0    1    0     1     1     1     1
#>  [14,]    0    1   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [15,]    0    0    0    0    0    0    0    0    0     0    NA    NA    NA
#>  [16,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [18,]    1    1    1    0    1    1    1    1    1     1     1     1     1
#>  [19,]    0    0    0    0    0    0    0    0   NA    NA    NA    NA    NA
#>  [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [22,]    1    1    1    1    0    1    1    1    1     1     1     1     1
#>  [23,]    0    0    0    0    0    0    0   NA   NA    NA    NA    NA    NA
#>  [24,]    1    1    1    0    1    1    1    1    1     1    NA    NA    NA
#>  [25,]    1    1    1    0    0    1    0    1    1     1     1    NA    NA
#>  [26,]    1    1    1    1    1    1    1    1    1     1     1    NA    NA
#>  [27,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [28,]    1    1    1    1    1   NA   NA   NA   NA    NA    NA    NA    NA
#>  [29,]    1    1    1    1    1    1    1    1    0     0     0    NA    NA
#>  [30,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [31,]    0    0    0    0    0    0    0    0    0     0    NA    NA    NA
#>  [32,]    1    1    1    1    1    1    1    1    1     1     1     0     1
#>  [33,]    1    1   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [34,]    1    1    1    1    1    1    1    0    1     1     1     1     1
#>  [35,]    1    1    0    1    0    1    1   NA   NA    NA    NA    NA    NA
#>  [36,]    1    0    1    0    1    0    1    1    0     0     1     1     0
#>  [37,]    1    0    1    1    1    1    1    1   NA    NA    NA    NA    NA
#>  [38,]    1    1    1    1    1    1    1    1    1     1     0     1     1
#>  [39,]    0    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [40,]    1    1    0    1    1    1    1    1    0     1     0    NA    NA
#>  [41,]    1    1    1    1    1    1    1    0    1    NA    NA    NA    NA
#>  [42,]    1    1   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [43,]    1    0    1    1    1    1    1    1    0     0     1     1     1
#>  [44,]    1    1    0    1    1    1    1    1    0     0     1     1     1
#>  [45,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [46,]    1    0    0    1    1    1   NA   NA   NA    NA    NA    NA    NA
#>  [47,]    1    1    1    0    0    1    1    0    1     1    NA    NA    NA
#>  [48,]    0    0    0    0    0    0    0   NA   NA    NA    NA    NA    NA
#>  [49,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>  [50,]    0    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [51,]    0    0    0    0    0    0    0    0    0     0     0     0    NA
#>  [52,]    0    0    0    0    0   NA   NA   NA   NA    NA    NA    NA    NA
#>  [53,]    1    0    1    1    1    1    1    1    1     0     1     1     1
#>  [54,]    1    1    1    1    1    0    0    1    1     1     1     1     1
#>  [55,]    0    0    0    0    0   NA   NA   NA   NA    NA    NA    NA    NA
#>  [56,]    1    1    1   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [57,]    1    1    1    0    1    0    1    1    1     1    NA    NA    NA
#>  [58,]    1    1    1    1    1   NA   NA   NA   NA    NA    NA    NA    NA
#>  [59,]    0    0    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [60,]    1    0    1    1    1    1    1    1    1     1     0    NA    NA
#>  [61,]    1    1    1    1    1    1    1   NA   NA    NA    NA    NA    NA
#>  [62,]    0    0    0    0    0   NA   NA   NA   NA    NA    NA    NA    NA
#>  [63,]    0    0    0    0    0    0    0    0    0     0    NA    NA    NA
#>  [64,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [65,]    1    1    0    1    1    1    1    1    1     1     1     0     1
#>  [66,]    0    0    1   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [67,]    1    1    1    1    1    0    1    1    1     1     0    NA    NA
#>  [68,]    1    0    1    1    1    1    1   NA   NA    NA    NA    NA    NA
#>  [69,]    0    1    1    1    1    1   NA   NA   NA    NA    NA    NA    NA
#>  [70,]    0    1    1    1    1    1    1    1    1     1     1     1     1
#>  [71,]    1    0    1    1    1    0    1    1    1     1     1     1    NA
#>  [72,]    1    1    1    1    1    1    1    1    1    NA    NA    NA    NA
#>  [73,]    0    1    1    1    0    1    0    1    1     1     1     1    NA
#>  [74,]    1    1    1    0    1    1    1    0    1     1     1     1     1
#>  [75,]    1    1   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [76,]    0    0    0    0    0    0    0    0    0    NA    NA    NA    NA
#>  [77,]    1    0    1    1    1    1    0    1    1     1     1     1     0
#>  [78,]    0    0    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [79,]    0    1    1    1    1    1    1   NA   NA    NA    NA    NA    NA
#>  [80,]    0    0    0    1    0    1    1    1    0     1     0     0    NA
#>  [81,]    1    1    1    0    0    1    1    1    1     0    NA    NA    NA
#>  [82,]    1    1    1    0    1    0    1    1    1     1    NA    NA    NA
#>  [83,]    0    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [84,]    1    1    1   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [85,]    1    1   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [86,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [87,]    1    0    1    1    1    0    1    1    0     1     1     1    NA
#>  [88,]    1    1    1    0    1    1    1    1    0     1     0     1     0
#>  [89,]    0    1    0    0    1    1    1    1    0     0     1     1     1
#>  [90,]    1    1    1    0    1    1    1    1    1     0     1     1     1
#>  [91,]    1    1    1    1    1    1    1    1    1     0     1     1     1
#>  [92,]    1    1    0    1    1    1    1    1    1     1     1     1    NA
#>  [93,]    1    1    1    0    1    1    1    1    1     1    NA    NA    NA
#>  [94,]    0    0    0    0    0    0    0   NA   NA    NA    NA    NA    NA
#>  [95,]    1    1    1    1    1    0    1    1    1     1     1     1    NA
#>  [96,]    1    1    1    1    1    1    1    1    1     1     1    NA    NA
#>  [97,]    0    0    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [98,]    1    1    0    1    0    1    1    1    1     1    NA    NA    NA
#>  [99,]    1    1    1    1    0    0    1    0    1     1     1     1     1
#> [100,]    0    0    1   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>        [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>   [1,]     0     1     1    NA    NA    NA    NA
#>   [2,]     0     1     1     1     1     1    NA
#>   [3,]     0     0     0     0     0     0     0
#>   [4,]    NA    NA    NA    NA    NA    NA    NA
#>   [5,]     0     1     0     1     1    NA    NA
#>   [6,]     1     0     1    NA    NA    NA    NA
#>   [7,]     1     1     1     1    NA    NA    NA
#>   [8,]     1     1     1     1     1    NA    NA
#>   [9,]    NA    NA    NA    NA    NA    NA    NA
#>  [10,]     0     0     0     0     0     0     0
#>  [11,]    NA    NA    NA    NA    NA    NA    NA
#>  [12,]    NA    NA    NA    NA    NA    NA    NA
#>  [13,]    NA    NA    NA    NA    NA    NA    NA
#>  [14,]    NA    NA    NA    NA    NA    NA    NA
#>  [15,]    NA    NA    NA    NA    NA    NA    NA
#>  [16,]     0     0    NA    NA    NA    NA    NA
#>  [17,]    NA    NA    NA    NA    NA    NA    NA
#>  [18,]     1    NA    NA    NA    NA    NA    NA
#>  [19,]    NA    NA    NA    NA    NA    NA    NA
#>  [20,]    NA    NA    NA    NA    NA    NA    NA
#>  [21,]    NA    NA    NA    NA    NA    NA    NA
#>  [22,]     1     1    NA    NA    NA    NA    NA
#>  [23,]    NA    NA    NA    NA    NA    NA    NA
#>  [24,]    NA    NA    NA    NA    NA    NA    NA
#>  [25,]    NA    NA    NA    NA    NA    NA    NA
#>  [26,]    NA    NA    NA    NA    NA    NA    NA
#>  [27,]     0     0     0     0    NA    NA    NA
#>  [28,]    NA    NA    NA    NA    NA    NA    NA
#>  [29,]    NA    NA    NA    NA    NA    NA    NA
#>  [30,]     0     0     0    NA    NA    NA    NA
#>  [31,]    NA    NA    NA    NA    NA    NA    NA
#>  [32,]     1     0    NA    NA    NA    NA    NA
#>  [33,]    NA    NA    NA    NA    NA    NA    NA
#>  [34,]     1     0     1     1     1     1     1
#>  [35,]    NA    NA    NA    NA    NA    NA    NA
#>  [36,]    NA    NA    NA    NA    NA    NA    NA
#>  [37,]    NA    NA    NA    NA    NA    NA    NA
#>  [38,]     0     1    NA    NA    NA    NA    NA
#>  [39,]    NA    NA    NA    NA    NA    NA    NA
#>  [40,]    NA    NA    NA    NA    NA    NA    NA
#>  [41,]    NA    NA    NA    NA    NA    NA    NA
#>  [42,]    NA    NA    NA    NA    NA    NA    NA
#>  [43,]     1     1     1     1     0    NA    NA
#>  [44,]     1     1     0     1    NA    NA    NA
#>  [45,]     0     0    NA    NA    NA    NA    NA
#>  [46,]    NA    NA    NA    NA    NA    NA    NA
#>  [47,]    NA    NA    NA    NA    NA    NA    NA
#>  [48,]    NA    NA    NA    NA    NA    NA    NA
#>  [49,]     0     0     0     0     0     0    NA
#>  [50,]    NA    NA    NA    NA    NA    NA    NA
#>  [51,]    NA    NA    NA    NA    NA    NA    NA
#>  [52,]    NA    NA    NA    NA    NA    NA    NA
#>  [53,]     1     1     1     0     1     1     1
#>  [54,]     1     1    NA    NA    NA    NA    NA
#>  [55,]    NA    NA    NA    NA    NA    NA    NA
#>  [56,]    NA    NA    NA    NA    NA    NA    NA
#>  [57,]    NA    NA    NA    NA    NA    NA    NA
#>  [58,]    NA    NA    NA    NA    NA    NA    NA
#>  [59,]    NA    NA    NA    NA    NA    NA    NA
#>  [60,]    NA    NA    NA    NA    NA    NA    NA
#>  [61,]    NA    NA    NA    NA    NA    NA    NA
#>  [62,]    NA    NA    NA    NA    NA    NA    NA
#>  [63,]    NA    NA    NA    NA    NA    NA    NA
#>  [64,]    NA    NA    NA    NA    NA    NA    NA
#>  [65,]     1     1     1     0     1     1     1
#>  [66,]    NA    NA    NA    NA    NA    NA    NA
#>  [67,]    NA    NA    NA    NA    NA    NA    NA
#>  [68,]    NA    NA    NA    NA    NA    NA    NA
#>  [69,]    NA    NA    NA    NA    NA    NA    NA
#>  [70,]     1    NA    NA    NA    NA    NA    NA
#>  [71,]    NA    NA    NA    NA    NA    NA    NA
#>  [72,]    NA    NA    NA    NA    NA    NA    NA
#>  [73,]    NA    NA    NA    NA    NA    NA    NA
#>  [74,]     0     1     1     0    NA    NA    NA
#>  [75,]    NA    NA    NA    NA    NA    NA    NA
#>  [76,]    NA    NA    NA    NA    NA    NA    NA
#>  [77,]    NA    NA    NA    NA    NA    NA    NA
#>  [78,]    NA    NA    NA    NA    NA    NA    NA
#>  [79,]    NA    NA    NA    NA    NA    NA    NA
#>  [80,]    NA    NA    NA    NA    NA    NA    NA
#>  [81,]    NA    NA    NA    NA    NA    NA    NA
#>  [82,]    NA    NA    NA    NA    NA    NA    NA
#>  [83,]    NA    NA    NA    NA    NA    NA    NA
#>  [84,]    NA    NA    NA    NA    NA    NA    NA
#>  [85,]    NA    NA    NA    NA    NA    NA    NA
#>  [86,]    NA    NA    NA    NA    NA    NA    NA
#>  [87,]    NA    NA    NA    NA    NA    NA    NA
#>  [88,]     0     1     1     1    NA    NA    NA
#>  [89,]     1     1    NA    NA    NA    NA    NA
#>  [90,]    NA    NA    NA    NA    NA    NA    NA
#>  [91,]    NA    NA    NA    NA    NA    NA    NA
#>  [92,]    NA    NA    NA    NA    NA    NA    NA
#>  [93,]    NA    NA    NA    NA    NA    NA    NA
#>  [94,]    NA    NA    NA    NA    NA    NA    NA
#>  [95,]    NA    NA    NA    NA    NA    NA    NA
#>  [96,]    NA    NA    NA    NA    NA    NA    NA
#>  [97,]    NA    NA    NA    NA    NA    NA    NA
#>  [98,]    NA    NA    NA    NA    NA    NA    NA
#>  [99,]    NA    NA    NA    NA    NA    NA    NA
#> [100,]    NA    NA    NA    NA    NA    NA    NA